move the DateTimeEncoder to the span module where it might be spotted for reuse later
